REFACTOR: Rename Formatter -> LogFormatter
